pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "volume_price_trade"
version = "0.1.0"
description = "Intraday ML backtesting with Volume Profile, VPA, ICT on SP500 1m data."
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.2",
  "numpy>=1.26",
  "pyarrow>=15.0",
  "gcsfs>=2024.6.0",
  "pytz",
  "python-dateutil",
  "scikit-learn>=1.4",
  "xgboost>=2.0",
  "lightgbm>=4.3",
  "matplotlib>=3.8",
  "plotly>=5.20",
  "pyyaml>=6.0.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=4.1",
  "mypy>=1.10",
  "ruff>=0.5.0",
  "types-PyYAML",
  "types-python-dateutil",
  "types-requests"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E","F","I","W","B","UP"]
ignore = []

[tool.mypy]
python_version = "3.10"
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "UP"]
ignore = []
